% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_livimal.R
\name{add_livimal}
\alias{add_livimal}
\title{Add a livimals silhouette to a ggplot!}
\usage{
add_livimal(
  animal,
  x = 1,
  y = 1,
  color = "black",
  mapping = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{animal}{Name of the image file (without .png)}

\item{x}{Optional x-position. Used only when no aes() is supplied.}

\item{y}{Optional y-position. Used only when no aes() is supplied.}

\item{color}{Optional color to apply to the silhouette (default "black")}

\item{mapping}{Optional aes() for x, y, or image.}

\item{size}{Size of the image (0.1 is small, 0.5 is big)}
}
\value{
A ggplot2 layer
}
\description{
Add a livimals silhouette to a ggplot!
}
