% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_livimals.R
\name{col_livimal}
\alias{col_livimal}
\title{Add a livimals silhouette to a ggplot!}
\usage{
col_livimal(
  animal,
  x = 1,
  y = 1,
  color = "black",
  mapping = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{animal}{Name of the image file (without .png)}

\item{x}{Optional x-position. Used only when no aes() is supplied.}

\item{y}{Optional y-position. Used only when no aes() is supplied.}

\item{color}{Optional color to apply to the silhouette (default "black")}

\item{mapping}{Optional aes() for x, y, or image.}

\item{data}{A dataframe}

\item{...}{Additional arguments passed to geom_image()}
}
\value{
A ggplot2 layer
}
\description{
Add a livimals silhouette to a ggplot!
}
