% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_livimals.R
\name{list_livimals}
\alias{list_livimals}
\title{List all available livimals animals}
\usage{
list_livimals()
}
\value{
A character vector of animal names (without file extensions)
}
\description{
List all available livimals animals
}
