% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_livimal.R
\name{preview_livimal}
\alias{preview_livimal}
\title{Preview an animal}
\usage{
preview_livimal(animal)
}
\arguments{
\item{animal}{Character string. The name of the animal to preview.}
}
\value{
Displays the image in the R plotting window.
}
\description{
This function displays the silhouette of a single animal.
}
